requirements.txt
astroid==1.6.1
autoenv==1.0.0
certifi==2018.1.18
chardet==3.0.4
coverage==4.4.2
defusedxml==0.5.0
dj-database-url==0.4.2
dj-static==0.0.6
Django==1.11.9
django-coverage-plugin==1.5.0
django-debug-toolbar==1.9.1
django-toolbelt==0.0.1
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
gunicorn==19.7.1
idna==2.6
isort==4.3.0
lazy-object-proxy==1.3.1
mccabe==0.6.1
nose==1.3.7
oauthlib==2.0.6
psycopg2==2.7.3.2
PyJWT==1.5.3
pylint==1.8.2
pylint-django==0.9.0
pylint-plugin-utils==0.2.6
python3-openid==3.1.0
pytz==2017.3
requests==2.18.4
requests-oauthlib==0.8.0
six==1.11.0
social-auth-app-django==2.1.0
social-auth-core==1.6.0
sqlparse==0.2.4
static3==0.7.0
urllib3==1.22
wrapt==1.10.11
